requirements.txt
# Core AutoGen
autogen-agentchat>=0.7.5
autogen-ext>=0.7.5
autogen-core>=0.7.5

# MCP Support
mcp>=0.1.0
httpx>=0.27.0

# LLM Support
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio>=3.4.3
colorama>=0.4.6